JPAEmitter OK + header reorganization
