+++ libs/JSystem/JParticle/JPABaseShape.cpp
@@ -5,6 +5,7 @@
 
 #include "JSystem/JParticle/JPABaseShape.h"
 #include "JSystem/JParticle/JPAParticle.h"
+#include "JSystem/JParticle/JPAEmitter.h"
 #include "JSystem/JParticle/JPAResource.h"
 #include "JSystem/JParticle/JPAResourceManager.h"
 #include "dolphin/mtx/mtxvec.h"
