+++ libs/JSystem/JParticle/JPADynamicsBlock.cpp
@@ -5,7 +5,7 @@
 
 #include "JSystem/JParticle/JPADynamicsBlock.h"
 #include "JSystem/JMath/JMATrigonometric.h"
-#include "JSystem/JParticle/JPAParticle.h"
+#include "JSystem/JParticle/JPAEmitter.h"
 #include "dol2asm.h"
 #include "dolphin/types.h"
