+++ libs/JSystem/JParticle/JPAEmitterManager.cpp
@@ -5,7 +5,7 @@
 
 #include "JSystem/JParticle/JPAEmitterManager.h"
 #include "JSystem/JKernel/JKRHeap.h"
-#include "JSystem/JParticle/JPAParticle.h"
+#include "JSystem/JParticle/JPAEmitter.h"
 #include "JSystem/JParticle/JPAResourceManager.h"
 #include "dol2asm.h"
 #include "dolphin/gx/GX.h"
