+++ libs/JSystem/JParticle/JPAParticle.cpp
@@ -4,6 +4,8 @@
 //
 
 #include "JSystem/JParticle/JPAParticle.h"
+#include "JSystem/JParticle/JPAEmitter.h"
+#include "JSystem/JParticle/JPAEmitterManager.h"
 #include "dolphin/mtx/mtxvec.h"
 #include "dol2asm.h"
 #include "dolphin/types.h"
