+++ libs/JSystem/JParticle/JPAResource.cpp
@@ -7,6 +7,7 @@
 #include "JSystem/JParticle/JPABaseShape.h"
 #include "JSystem/JParticle/JPAMath.h"
 #include "JSystem/JParticle/JPAParticle.h"
+#include "JSystem/JParticle/JPAEmitter.h"
 #include "dol2asm.h"
 #include "dolphin/gx/GX.h"
 #include "dolphin/mtx/mtx.h"
@@ -762,7 +763,7 @@ void JPAResource::calcWorkData_d(JPAEmitterWorkData* work) {
     JPAGetXYZRotateMtx(work->mpEmtr->mLocalRot.x * 0xB6, work->mpEmtr->mLocalRot.y * 0xB6,
                        work->mpEmtr->mLocalRot.z * 0xB6, mtx);
     PSMTXConcat(work->mpEmtr->mGlobalRot, mtx, work->mGlobalRot);
-    PSMTXMultVecSR(work->mGlobalRot, &work->mpEmtr->mLocalDir, work->mGlobalEmtrDir);
+    PSMTXMultVecSR(work->mGlobalRot, (Vec*)&work->mpEmtr->mLocalDir, (Vec*)&work->mGlobalEmtrDir);
 }
 
 /* 80276A8C-80276A90 2713CC 0004+00 0/0 3/0 0/0 .text
