+++ src/d/particle/d_particle.cpp
@@ -5,6 +5,7 @@
 
 #include "d/particle/d_particle.h"
 #include "JSystem/JKernel/JKRExpHeap.h"
+#include "JSystem/JParticle/JPAEmitterManager.h"
 #include "MSL_C/stdio.h"
 #include "d/com/d_com_inf_game.h"
 #include "dol2asm.h"
